+++ bokeh/io/tests/test_doc.py
@@ -1,0 +1,71 @@
+#-----------------------------------------------------------------------------
+# Copyright (c) 2012 - 2017, Anaconda, Inc. All rights reserved.
+#
+# Powered by the Bokeh Development Team.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Boilerplate
+#----------------------------------------------------------------------------
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+import pytest ; pytest
+
+from bokeh.util.api import INTERNAL, PUBLIC ; INTERNAL, PUBLIC
+from bokeh.util.testing import verify_api ; verify_api
+
+#-----------------------------------------------------------------------------
+# Imports
+#-----------------------------------------------------------------------------
+
+# Standard library imports
+
+# External imports
+
+# Bokeh imports
+from bokeh.document import Document
+from bokeh.io.state import curstate
+
+# Module under test
+import bokeh.io.doc as bid
+
+#-----------------------------------------------------------------------------
+# API Definition
+#-----------------------------------------------------------------------------
+
+api = {
+
+    PUBLIC: (
+
+        ( 'curdoc', (1, 0, 0) ),
+
+    ), INTERNAL: (
+
+        ( 'set_curdoc', (1, 0, 0) ),
+    )
+
+}
+
+test_public_api, test_internal_api, test_all_declared, test_all_tested = verify_api(bid, api)
+
+#-----------------------------------------------------------------------------
+# Public API
+#-----------------------------------------------------------------------------
+
+def test_curdoc_from_curstate():
+    assert bid.curdoc() is curstate().document
+
+#-----------------------------------------------------------------------------
+# Internal API
+#-----------------------------------------------------------------------------
+
+def test_set_curdoc_sets_curstate():
+    d = Document()
+    bid.set_curdoc(d)
+    assert curstate().document is d
+
+#-----------------------------------------------------------------------------
+# Private API
+#-----------------------------------------------------------------------------
