+++ bokeh/plotting/__init__.py
@@ -10,7 +10,7 @@ from ..io import output_notebook; output_notebook
 from ..io import reset_output; reset_output
 from ..io import save; save
 from ..io import show; show
-from ..io import gridplot, GridSpec; gridplot, GridSpec
+from ..layouts import gridplot, GridSpec; gridplot, GridSpec
 
 from .figure import Figure; Figure
 from .figure import figure; figure
