+++ setup.cfg
@@ -21,6 +21,7 @@ python_files = *_tests.py *_test.py test_*.py
 selenium_exclude_debug = html logs
 implicit_marker = unit
 markers =
+    api: an api versioning check
     cross_browser: an integration test to run across multiple browsers on SauceLabs
     examples: an examples image-diff test
     integration: an integration test that runs on Saucelabs
